move to analysis circular buffer
